rename `label_handler` to `label_updator`
