+++ src/DAJIN2/core/clustering/__init__.py
@@ -1,3 +1,3 @@
 from DAJIN2.core.clustering.label_extractor import extract_labels
 from DAJIN2.core.clustering.appender import add_labels, add_readnum, add_percent
-from DAJIN2.core.clustering.label_handler import update_labels
+from DAJIN2.core.clustering.label_updator import update_labels
